+++ docs/changelog.md
@@ -38,7 +38,7 @@
 - new native section to detect native library availability.
 
 ## dependencies
-- remove `pastey`.
+- remove dependencies: `bytemuck`, `pastey`.
 - bump `portable-atomic-util` to 0.2.7.
 - bump `hashbrown` to 0.17.
 - bump `wide` to 1.3.